update string --> text
