+++ synthetic_data/dataset_generator.py
@@ -12,7 +12,7 @@ from synthetic_data.distinct_generators.categorical_generator import random_cate
 from synthetic_data.distinct_generators.datetime_generator import random_datetimes
 from synthetic_data.distinct_generators.float_generator import random_floats
 from synthetic_data.distinct_generators.int_generator import random_integers
-from synthetic_data.distinct_generators.string_generator import random_string
+from synthetic_data.distinct_generators.text_generator import random_text
 
 
 def convert_data_to_df(
@@ -60,7 +60,7 @@ def generate_dataset(
         "float": random_floats,
         "categorical": random_categorical,
         "datetime": random_datetimes,
-        "string": random_string,
+        "text": random_text,
     }
 
     if not columns_to_generate:
